requirements.txt
absl-py==2.0.0
addict==2.4.0
basicsr==1.4.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
fonttools==4.47.2
fsspec==2023.12.2
future==0.18.3
gfpgan==1.3.8
grpcio==1.60.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
protobuf==4.25.2
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
realesrgan==0.3.0
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sympy==1.12
tb-nightly==2.16.0a20240111
tensorboard-data-server==0.7.2
tf_keras-nightly==2.16.0.dev2024011110
thop==0.1.1.post2209072238
tifffile==2023.12.9
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
unet==0.7.7
urllib3==2.1.0
Werkzeug==3.0.1
yapf==0.40.2
zipp==3.17.0
